+++ src/lib/forms/Label.svelte
@@ -2,7 +2,7 @@
 	import classNames from 'classnames';
 
 	export let color: 'gray' | 'green' | 'red' | 'disabled' = 'gray';
-	export let defaultClass: string = 'text-sm font-medium';
+	export let defaultClass: string = 'text-sm font-medium block';
 
 	export let show: boolean = true; // helper for inheritance