+++ src/lib/forms/SimpleSearch.svelte
@@ -1,11 +1,16 @@
 <script lang="ts">
+	import { getContext } from 'svelte';
+
+	// tainted if put in component having its own background
+	let background: boolean = getContext('background');
+
 	export let id: string = '';
 	export let labelClass: string = 'sr-only';
 	export let iconClass: string = 'w-5 h-5 text-gray-500 dark:text-gray-400';
 	export let iconDivClass: string =
 		'flex absolute inset-y-0 left-0 items-center pl-3 pointer-events-none';
 	export let inputClass: string =
-		'bg-gray-50 border border-gray-300 text-gray-900 text-sm rounded-lg focus:ring-blue-500 focus:border-blue-500 block w-full pl-10 p-2.5  dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-blue-500 dark:focus:border-blue-500';
+		'bg-gray-50 border border-gray-300 text-gray-900 text-sm rounded-lg focus:ring-blue-500 focus:border-blue-500 block w-full pl-10 p-2.5 dark:placeholder-gray-400 dark:text-white dark:focus:ring-blue-500 dark:focus:border-blue-500';
 	export let btnClass: string =
 		'p-2.5 ml-2 text-sm font-medium text-white bg-blue-700 rounded-lg border border-blue-700 hover:bg-blue-800 focus:ring-4 focus:outline-none focus:ring-blue-300 dark:bg-blue-600 dark:hover:bg-blue-700 dark:focus:ring-blue-800';
 	export let placeholder: string = 'Search';
@@ -27,7 +32,15 @@
 				/></svg
 			>
 		</div>
-		<input {...$$restProps} type="text" {id} class={inputClass} {placeholder} />
+		<input
+			{...$$restProps}
+			type="text"
+			{id}
+			class="{background
+				? 'dark:bg-gray-600 dark:border-gray-500'
+				: ' dark:bg-gray-700 dark:border-gray-600'} {inputClass}"
+			{placeholder}
+		/>
 	</div>
 	<button type="submit" class={btnClass}
 		><svg