+++ src/routes/index.svelte
@@ -21,6 +21,7 @@
 		InputfieldBlock,
 		KbdBlock,
 		ListgroupBlock,
+		MegamenuBlock,
 		ModalBlock,
 		NavbarBlock,
 		PaginationBlock,
@@ -50,17 +51,18 @@
 <Responsive />
 
 <div class="relative bg-white overflow-hidden mt-8 lg:mb-8 mx-auto dark:bg-gray-900 2xl:px-8">
-	<div class="max-w-7xl mx-auto ">
+	<div class="max-w-full mx-auto ">
 		<div class="relative z-10 xl:pb-8 bg-transparent lg:max-w-2xl lg:w-full">
 			<div class="my-auto mx-auto max-w-7xl xl:pl-8 xl:pt-8">
 				<div class="sm:text-center lg:text-left">
-					<h1 class="text-4xl tracking-tight font-extrabold text-gray-900  dark:text-white">
+					<h1
+						class="max-w-md text-4xl tracking-tight font-extrabold text-gray-900  dark:text-white"
+					>
 						<span class="block xl:inline">Flowbite Svelte </span>
 						<span class="block text-indigo-600 dark:text-white xl:inline">UI Component</span>
 					</h1>
 					<p
-						class="mt-3 text-base text-gray-500 sm:mt-5 sm:text-lg 
-            lg:max-w-sm sm:max-w-xl sm:mx-auto md:mt-5 md:text-xl lg:mx-0 dark:text-white"
+						class="mt-3 text-base text-gray-500 sm:mt-5 sm:text-lg lg:max-w-sm sm:max-w-xl sm:mx-auto md:mt-5 md:text-xl lg:mx-0 dark:text-white"
 					>
 						Speed up your web development with Flowbite Svelte. Flowbite Svelte is an official
 						Flowbite component library for Svelte. All interactivities are handled by Svelte.
@@ -98,6 +100,7 @@
 	<InputfieldBlock />
 	<KbdBlock />
 	<ListgroupBlock />
+	<MegamenuBlock />
 	<ModalBlock />
 	<NavbarBlock />
 	<PaginationBlock />