+++ src/routes/moduleItems.js
@@ -1,206 +1,140 @@
 export const svelteflows = [
-  { href: "/pages/getting-started", name: "Getting Started" },
-  { href: "/pages/about", name: "About" },
-  { href: "/pages/types", name: "Types" },
-  { href: "/pages/how-to-contribute", name: "How to contribute" },
-  { href: "/pages/cli", name: "Flowbite-Svelte CLI" },
-  { href: "/pages/license", name: "License" },
+	{ href: '/pages/getting-started', name: 'Getting Started' },
+	{ href: '/pages/about', name: 'About' },
+	{ href: '/pages/types', name: 'Types' },
+	{ href: '/pages/how-to-contribute', name: 'How to contribute' },
+	{ href: '/pages/cli', name: 'Flowbite-Svelte CLI' },
+	{ href: '/pages/license', name: 'License' }
 ];
-
-export const accordions = [
-  { href: "/accordions/default", name: "Accordion" },
-  { href: "/accordions/icon-accordion", name: "Icon accordion" },
-];
-
-export const alerts = [
-  { href: "/alerts/", name: "Alert" },
-];
-
-export const badges = [
-  { href: "/badges/", name: "Badges" },
-];
-
-export const cards = [
-  { href: "/cards/card", name: "Card" },
-  { href: "/cards/ecommerce", name: "E-commerce" },
-  { href: "/cards/cta", name: "CTA" },
-  { href: "/cards/horizontal", name: "Horizontal" },
-  { href: "/cards/interactive", name: "Interactive" },
-  { href: "/cards/list", name: "List" },
-  { href: "/cards/signin", name: "SignIn" },
-];
-
-export const carousels = [
-  { href: "/carousels/default", name: "Default" },
-  { href: "/carousels/transition", name: "Transition" },
-]
 
 export const forms = [
-  { href: "/forms/checkbox", name: "Checkbox" },
-  { href: "/forms/file-input", name: "File input" },
-  { href: "/forms/floating-label", name: "Floating label" },
-  { href: "/forms/input-field", name: "Input" },
-  { href: "/forms/radio", name: "Radio" },
-  { href: "/forms/range", name: "Range" },
-  { href: "/forms/search-input", name: "Search" },
-  { href: "/forms/select", name: "Select" },
-  { href: "/forms/textarea", name: "Textarea" },
-  { href: "/forms/toggle", name: "Toggle" },
-]
-
-export const modals = [
-  { href: "/modals/small", name: "Small" },
-  { href: "/modals/medium", name: "Medium" },
-  { href: "/modals/large", name: "Large" },
-  { href: "/modals/extra-large", name: "Extra-large" },
-  { href: "/modals/signin", name: "Sign-in" },
-  { href: "/modals/all-modals", name: "All modals" },
+	{ href: '/forms/checkbox', name: 'Checkbox' },
+	{ href: '/forms/file-input', name: 'File input' },
+	{ href: '/forms/floating-label', name: 'Floating label' },
+	{ href: '/forms/input-field', name: 'Input' },
+	{ href: '/forms/radio', name: 'Radio' },
+	{ href: '/forms/range', name: 'Range' },
+	{ href: '/forms/search-input', name: 'Search' },
+	{ href: '/forms/select', name: 'Select' },
+	{ href: '/forms/textarea', name: 'Textarea' },
+	{ href: '/forms/toggle', name: 'Toggle' }
 ];
-
-export const ratings = [
-  { href: "/ratings/rating", name: "Rating" },
-  { href: "/ratings/advanced-rating", name: "Advanced rating" },
-  { href: "/ratings/score-rating", name: "Score rating" },
-]
-
-export const tabs = [
-  { href: "/tabs/default-tabs", name: "Default" },
-  { href: "/tabs/interactive-tabs", name: "Interactive" },
-  { href: "/tabs/interactive-tabs-2", name: "Interactive Tabs 2" },
-  { href: "/tabs/pilltabs", name: "Pilltabs" },
-  { href: "/tabs/full-width-tabs", name: "Full width tabs" },
-  { href: "/tabs/underline-tabs", name: "Underline tabs" },
-  { href: "/tabs/icon-tabs", name: "Icon tabs" },
-  { href: "/tabs/multiple-interactive-tabs", name: "Multiple Interactive" },
-];
-
-export const timelines = [
-  { href: "/timelines/default", name: "Default" },
-  { href: "/timelines/vertical", name: "Vertical" },
-  { href: "/timelines/horizontal", name: "Horizontal" },
-  { href: "/timelines/activity", name: "Activity log" },
-  { href: "/timelines/group", name: "Grouped"}
-]
-// export const tooltips = [
-//   { href: "/tooltips/default", name: "Default" },
-//   { href: "/tooltips/light", name: "Light tooltip" },
-// ];
 
 export const topMenus = [
-  { id:"1", href: "/", name: "Home" },
-  { id:"2", href: "/pages/about", name: "About" },
-  {
-    id:"3", href: "https://github.com/themesberg/flowbite-svelte",
-    name: "GitHub"
-  },
+	{ id: '1', href: '/', name: 'Home' },
+	{ id: '2', href: '/pages/about', name: 'About' },
+	{
+		id: '3',
+		href: 'https://github.com/themesberg/flowbite-svelte',
+		name: 'GitHub'
+	}
 ];
 
 export const footerLinks = [
-  {
-    parent: 'COMPONENTS',
-    children: [
-      { name: 'Accordion', href: '/accordions/' },
-      { name: 'Alert', href: '/alerts' },
-      { name: 'Avatar', href: '/avatar' },
-      { name: 'Badge', href: '/badges' },
-      { name: 'Breadcrumb', href: '/breadcrumbs' },
-      { name: 'Button', href: '/buttons' }
-    ]
-  },
-  {
-    parent: 'COMPONENTS',
-    children: [
-      { name: 'Button Group', href: '/button-groups' },
-      { name: 'Card', href: '/cards' },
-      { name: 'Carousel', href: '/carousels' },
-      { name: 'Dark', href: '/darkmode' },
-      { name: 'Dropdown', href: '/dropdowns' },
-      { name: 'Footer', href: '/footer' }
-    ]
-  },
-  {
-    parent: 'COMPONENTS',
-    children: [
-      { name: 'Form', href: '/forms' },
-      { name: 'Icon', href: '/icons' },
-      { name: 'List group', href: '/list-group' },
-      { name: 'Modal', href: '/modals' },
-      { name: 'Navbar', href: '/navbar' },
-      { name: 'Pagination', href: '/paginations' }
-    ]
-  },
-  {
-    parent: 'COMPONENTS',
-    children: [
-      { name: 'Progress bar', href: '/progressbars' },
-      { name: 'Rating', href: '/ratings' },
-      { name: 'Sidebar', href: '/sidebars' },
-      { name: 'Spinner', href: '/spinners' },
-      { name: 'Tables', href: '/tables' },
-      { name: 'Tab', href: '/tabs' }
-    ]
-  },
-  {
-    parent: 'COMPONENTS',
-    children: [
-      { name: 'Timeline', href: '/timelines' },
-      { name: 'Toast', href: '/toasts' },
-      { name: 'Tooltip', href: '/tooltips' }
-    ]
-  },
-  {
-    parent: 'DOCS',
-    children: [
-      { name: 'Getting Started', href: '/pages/getting-started' },
-      {
-        name: 'About',
-        href: '/pages/about'
-      },
-      {
-        name: 'Types',
-        href: '/pages/types'
-      },
-      {
-        name: 'How to contribute',
-        href: '/pages/how-to-contribute'
-      },
-      {
-        name: 'CLI',
-        href: '/pages/cli'
-      },
-      {
-        name: 'License',
-        href: '/pages/license'
-      }
-    ]
-  },
-  {
-    parent: 'RESOURCES',
-    children: [
-      {
-        name: 'Flowbite',
-        href: 'https://flowbite.com/docs/getting-started/introduction/'
-      },
-      {
-        name: 'Tailwind CSS',
-        href: 'https://tailwindcss.com/docs/installation'
-      }
-    ]
-  },
-  {
-    parent: 'FOLLOW US',
-    children: [
-      { name: 'Github', href: 'https://github.com/themesberg/flowbite-svelte' },
-      {
-        name: 'Twitter',
-        href: 'https://twitter.com/shinokada'
-      }
-    ]
-  },
-  {
-    parent: 'LEGAL',
-    children: [{ name: 'License', href: 'license' }]
-  }
-]
-
-
+	{
+		parent: 'COMPONENTS',
+		children: [
+			{ name: 'Accordion', href: '/accordions/' },
+			{ name: 'Alert', href: '/alerts' },
+			{ name: 'Avatar', href: '/avatar' },
+			{ name: 'Badge', href: '/badges' },
+			{ name: 'Breadcrumb', href: '/breadcrumbs' },
+			{ name: 'Button', href: '/buttons' }
+		]
+	},
+	{
+		parent: 'COMPONENTS',
+		children: [
+			{ name: 'Button Group', href: '/button-groups' },
+			{ name: 'Card', href: '/cards' },
+			{ name: 'Carousel', href: '/carousels' },
+			{ name: 'Dark', href: '/darkmode' },
+			{ name: 'Dropdown', href: '/dropdowns' },
+			{ name: 'Footer', href: '/footer' }
+		]
+	},
+	{
+		parent: 'COMPONENTS',
+		children: [
+			{ name: 'Form', href: '/forms' },
+			{ name: 'Icon', href: '/icons' },
+			{ name: 'List group', href: '/list-group' },
+			{ name: 'Modal', href: '/modals' },
+			{ name: 'Mega menu', href: '/megamenu' },
+			{ name: 'Navbar', href: '/navbar' },
+			{ name: 'Pagination', href: '/paginations' }
+		]
+	},
+	{
+		parent: 'COMPONENTS',
+		children: [
+			{ name: 'Progress bar', href: '/progressbars' },
+			{ name: 'Rating', href: '/ratings' },
+			{ name: 'Sidebar', href: '/sidebars' },
+			{ name: 'Spinner', href: '/spinners' },
+			{ name: 'Tables', href: '/tables' },
+			{ name: 'Tab', href: '/tabs' }
+		]
+	},
+	{
+		parent: 'COMPONENTS',
+		children: [
+			{ name: 'Timeline', href: '/timelines' },
+			{ name: 'Toast', href: '/toasts' },
+			{ name: 'Tooltip', href: '/tooltips' }
+		]
+	},
+	{
+		parent: 'DOCS',
+		children: [
+			{ name: 'Getting Started', href: '/pages/getting-started' },
+			{
+				name: 'About',
+				href: '/pages/about'
+			},
+			{
+				name: 'Types',
+				href: '/pages/types'
+			},
+			{
+				name: 'How to contribute',
+				href: '/pages/how-to-contribute'
+			},
+			{
+				name: 'CLI',
+				href: '/pages/cli'
+			},
+			{
+				name: 'License',
+				href: '/pages/license'
+			}
+		]
+	},
+	{
+		parent: 'RESOURCES',
+		children: [
+			{
+				name: 'Flowbite',
+				href: 'https://flowbite.com/docs/getting-started/introduction/'
+			},
+			{
+				name: 'Tailwind CSS',
+				href: 'https://tailwindcss.com/docs/installation'
+			}
+		]
+	},
+	{
+		parent: 'FOLLOW US',
+		children: [
+			{ name: 'Github', href: 'https://github.com/themesberg/flowbite-svelte' },
+			{
+				name: 'Twitter',
+				href: 'https://twitter.com/shinokada'
+			}
+		]
+	},
+	{
+		parent: 'LEGAL',
+		children: [{ name: 'License', href: 'license' }]
+	}
+];